% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_topic_model_report.R
\name{create_topic_model_report}
\alias{create_topic_model_report}
\title{Create Standard Results Report for a Structural Topic Model}
\usage{
create_topic_model_report(model, report_dir, cov_dataset, raw_texts,
  diff_cov = NULL, diff_cov.value1 = NULL, diff_cov.value2 = NULL,
  representative_docs = TRUE)
}
\arguments{
\item{model}{An object of class STM.}

\item{report_dir}{Directory in which to save the report.}

\item{cov_dataset}{A data frame containing any document-level covariates. Must contain a variable named "year".}

\item{raw_texts}{A character vector containing the texts analyzed in the topic model in a human-readable format, to output most representative documents for each topic.}

\item{diff_cov}{The name of a (factor or character) variable in \code{cov_dataset} with which to produce a difference plot. If NULL, no plot will be produced.}

\item{diff_cov.value1}{The first level of \code{diff_cov} to be used to create a difference plot.}

\item{diff_cov.value2}{The second level of \code{diff_cov} to be used to create a difference plot.}

\item{representative_docs}{If \code{TRUE} (the default), save full texts of most representative documents. If \code{FALSE}, save only a list of titles of the most representative documents.}
}
\description{
Create standard results report for a structural topic model, including FREX and PROB words, topic prevalence plot, topic time trend plots, and most representative documents.
}
