% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.table.R
\name{data.table}
\alias{data.table}
\title{data.table package}
\description{
This package imports all functions from the data.table package so it can
refer to them more easily. For more information, see
\code{\link[data.table]{data.table-package}}.
}
