% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frex.R
\name{get_frex}
\alias{get_frex}
\title{Get FREX words}
\usage{
get_frex(prob_mtrx, group_num, group_membership = NULL, wt = 0.5,
  values = FALSE)
}
\arguments{
\item{prob_mtrx}{A matrix with (named) columns for words, and rows for either documents or document groups. If each row represents a document, \code{group_membership} must be provided.}

\item{group_num}{An integer corresponding to the document group of interest.}

\item{group_membership}{An integer vector specifying the group membership of documents in \code{prob_mtrx}.}

\item{wt}{A value between 0 and 1 specifying the weight to be given to word exclusivity in calculating FREX.}

\item{values}{A logical specifying whether the function should return the actual FREX values or just words in decreasing FREX order.}
}
\description{
Get frequent and exclusive words for a group of documents, relative to other groups, based on a probability matrix of word occurrence.
}
