% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_words.R
\name{get_long_words}
\alias{get_long_words}
\title{Get Long Words}
\usage{
get_long_words(x, length_threshold = 20)
}
\arguments{
\item{x}{An object inheriting from Corpus, a list of character vectors, or a character vector.}

\item{length_threshold}{An integer. Words of length strictly greater than this threshold will be returned.}
}
\value{
A character vector of words with length greater than \code{length_threshold}.
}
\description{
This function takes a Corpus or vector of documents and returns words with length greater than a threshold.
}
\examples{
docs <- c("What are you doing down there, crashy?", "I'm real good at lying to Lois.")
long_words <- get_long_words(docs, length_threshold=4)
}
\keyword{text}
