% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topic_prevalence_by_covariate.R
\name{get_topic_prevalence_by_covariate}
\alias{get_topic_prevalence_by_covariate}
\title{Topic Prevalence by Covariate}
\usage{
get_topic_prevalence_by_covariate(stm_model, by_what, data)
}
\arguments{
\item{stm_model}{An object of class STM.}
}
\value{
A data table whose first column contains the levels of the covariate and subsequent columns give the average topic prevalence of each topic in documents at each level of the covariate.
}
\description{
This function returns a data table with the mean topic prevalence for each level of a character or factor covariate.
}
\keyword{text}
