% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_word.R
\name{interpret_word}
\alias{interpret_word}
\title{Interpret word}
\usage{
interpret_word(word, tkd_texts, doc_term_mtrx, stemmed_texts = NULL,
  summary = TRUE, custom_stem = NULL)
}
\arguments{
\item{word}{The word to be interpreted.}

\item{tkd_texts}{The tokenized texts used to create the document-term matrix.}

\item{doc_term_mtrx}{A document-term matrix summarizing the content of \code{tkd_texts}.}

\item{stemmed_texts}{If the texts were stemmed, this should be the stemmed and tokenized texts, otherwise \code{NULL}.}

\item{summary}{Logical: Should only a summary be printed, or all information returned?}

\item{custom_stem}{A character containing the regex pattern to match if \code{word} is a custom stem.}
}
\description{
Outputs information on the frequency and context of a word in documents analyzed by a topic model
}
