% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdf_summary_report.R
\name{make_pdf_summary_report}
\alias{make_pdf_summary_report}
\title{Create Standard Summary Report for STM Results}
\usage{
make_pdf_summary_report(stm_input_file, stm_results_dir, report_title,
  geogs = c("EU", "UN", "US"))
}
\arguments{
\item{stm_input_file}{Path to file containing objects used as input to the STM analysis.}

\item{stm_results_dir}{Directory containing results from STM analysis. The .Rdata file containing model output should have the same name as this directory.}

\item{report_title}{Title for the summary report. This is also used to name the output file, and thus should not contain special characters.}

\item{geogs}{Which geographies to include in the report.}
}
\description{
Create standard summary report for STM results.
}
