% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_csv.R
\name{make_summary_csv}
\alias{make_summary_csv}
\title{Save summary csv}
\usage{
make_summary_csv(stm_model, filename, type = c("frex", "prob"),
  n_words = 150)
}
\arguments{
\item{stm_model}{Output from a call to \code{stm}.}

\item{filename}{Path to write the file to.}

\item{type}{frex or prob.}

\item{n_words}{How many words per topic?}
}
\description{
Save a csv file with the highest FREX or probability words for each topic in a structural topic model
}
