% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_text_table.R
\name{make_text_table}
\alias{make_text_table}
\title{Make text table}
\usage{
make_text_table(col_list, headers, which_pretty = NULL, digits = rep(0,
  length(col_list)))
}
\arguments{
\item{col_list}{List of vectors containing the data for the columns of the table to be output.}

\item{headers}{Character vector of column headers.}

\item{which_pretty}{Integer vector containing column indices of columns to which \code{prettyNum} should be applied.}

\item{digits}{Integer vector giving the number of digits to be displayed after the decimal for each column. This value will be ignored for non-numeric columns.}
}
\description{
Make a nicely-formatted plain text table.
}
