% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_topic_network_graph.R
\name{make_topic_network_graph}
\alias{make_topic_network_graph}
\title{Make topic network graph}
\usage{
make_topic_network_graph(stm_results, stm_input_file,
  output_file = file.path(stm_results, "topic_network.png"))
}
\arguments{
\item{stm_results}{Path to directory containing STM results. This directory should have the same name as the .Rdata file containing the results, and the same name as the results object contained in that file.}

\item{stm_input_file}{Path to .Rdata file containing R object used as input for the STM analysis. This object should be named "stm_input".}

\item{output_file}{Path to .png file in which to save network graph.}
}
\description{
Make topic network graph from STM results.
}
