% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_color_coded_pdf.R
\name{print_color_coded_pdf}
\alias{print_color_coded_pdf}
\title{Save a color-coded PDF}
\usage{
print_color_coded_pdf(raw_text, frex, filename, stemmed = TRUE,
  colors = c("greyscale", "heat"), pctile_cuts = c(95, 98, 99))
}
\arguments{
\item{raw_text}{A character vector containing the text in human-readable format (with punctuation, case, etc)}

\item{frex}{A character vector of words sorted in decreasing order of FREX score.}

\item{filename}{Filepath to which output should be saved.}

\item{stemmed}{Are the words in \code{frex} stemmed?}

\item{colors}{One of \code{c("heat", "greyscale")} indicating color scheme for color coding.}

\item{pctile_cuts}{An integer of length 3 containing three numbers between 1 and 99 indicating the percentile cut-points for color-coding.}
}
\description{
Write text to a PDF with words color-coded according to FREX scores.
}
