% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_punctuation.R
\name{remove_punctuation}
\alias{remove_punctuation}
\title{Faster Punctuation Removal Courtesy of \code{stringr}}
\usage{
remove_punctuation(x, punct_pattern = "[^\\\\w ]|_")
}
\arguments{
\item{x}{A character vector.}

\item{punct_pattern}{Regular expression pattern to match for punctuation. Defaults to \code{"[^\\w ]|_"}, which matches any character that is not alphanumeric or a space, including underscores.}
}
\value{
A character vector with punctuation removed.
}
\description{
Uses \code{stringr}'s \code{str_replace_all} to remove punctuation.
}
\examples{
docs <- c("What are you doing down there, crashy?", "I'm real good at lying to Lois.")
docs <- remove_punctuation(docs)
}
\keyword{text}
