% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdize_for_comparison.R
\name{stdize_for_comparison}
\alias{stdize_for_comparison}
\title{Standardize to Find Duplicates}
\usage{
stdize_for_comparison(text)
}
\arguments{
\item{text}{A character vector}
}
\value{
A character vector with entries standardized.
}
\description{
This function standardizes case and removes all non-alphanumeric characters from a document to facilitate comparison and identification of effective duplicates.
}
\keyword{text}
